@@ -1,8 +1,38 @@
 {
   "name": "didyoumean2",
   "version": "1.3.0",
+  "license": "MIT",
+  "main": "src/index.js",
   "engines": {
     "node": ">=4.2"
+  },
+  "sideEffects": false,
+  "scripts": {
+    "ignore-sync": "ignore-sync",
+    "lint": "eslint .",
+    "test": "cross-env NODE_ENV=test jest --config .jestrc.json"
+  },
+  "dependencies": {
+    "leven": "^2.0.0",
+    "lodash.deburr": "^4.1.0",
+    "ramda": "^0.25.0"
+  },
+  "devDependencies": {
+    "@commitlint/cli": "^7.2.1",
+    "@commitlint/config-conventional": "^7.1.2",
+    "babel-core": "^6.26.0",
+    "babel-jest": "^23.6.0",
+    "babel-preset-env": "^1.6.1",
+    "cross-env": "^5.0.0",
+    "eslint": "^5.8.0",
+    "eslint-config-foray1010": "^6.2.0",
+    "husky": "^1.1.2",
+    "ignore-sync": "^1.0.0",
+    "jest": "^23.6.0",
+    "lint-staged": "^7.0.0",
+    "markdownlint-cli": "^0.13.0",
+    "prettier-eslint-cli": "^4.4.0",
+    "seamless-immutable": "^7.0.0"
   },
   "description": "a library for matching human-quality input to a list of potential matches using the Levenshtein distance algorithm",
   "keywords": [
@@ -28,30 +58,8 @@
     "text",
     "word"
   ],
-  "license": "MIT",
   "author": "foray1010",
   "homepage": "https://github.com/foray1010/didyoumean2",
   "repository": "foray1010/didyoumean2",
-  "bugs": "https://github.com/foray1010/didyoumean2/issues",
-  "main": "src/index.js",
-  "scripts": {
-    "lint": "eslint .",
-    "np": "np",
-    "precommit": "yarn run lint",
-    "test": "cross-env NODE_ENV=test jest --config .jestrc.json"
-  },
-  "dependencies": {
-    "leven": "^2.0.0",
-    "lodash": "^4.17.2"
-  },
-  "devDependencies": {
-    "cross-env": "^5.0.0",
-    "eslint": "^4.8.0",
-    "eslint-config-foray1010": "^6.2.0",
-    "husky": "^0.14.0",
-    "jest": "^21.0.0",
-    "np": "^2.11.0",
-    "pkg-dir": "^2.0.0",
-    "seamless-immutable": "^7.0.0"
-  }
+  "bugs": "https://github.com/foray1010/didyoumean2/issues"
 }
