@@ -1,75 +1,72 @@
-'use strict'
+import * as R from 'ramda'
 
-const returnTypeEnums = require('../enums/returnTypeEnums')
-const simpleSchema = require('./simpleSchema')
-const thresholdTypeEnums = require('../enums/thresholdTypeEnums')
-
-const EDIT_DISTANCE = thresholdTypeEnums.EDIT_DISTANCE
-const SIMILARITY = thresholdTypeEnums.SIMILARITY
+import returnTypeEnums from '../enums/returnTypeEnums.json'
+import thresholdTypeEnums from '../enums/thresholdTypeEnums.json'
+import simpleSchema from './simpleSchema'
 
 /**
  * Run `simpleSchema` on `options`
- * @param {null|Object|undefined} matchList - An options that allows you to modify the behavior
+ * @param {null|Object|undefined} options - An options that allows you to modify the behavior
  * @return {Object} options with default values
  */
-function runOptionsSchema(options) {
-  options = simpleSchema(options, {
-    type: 'object',
-    defaultValue: {}
-  })
-
-  options.caseSensitive = simpleSchema(options.caseSensitive, {
-    type: 'boolean',
-    defaultValue: false
-  })
-
-  options.deburr = simpleSchema(options.deburr, {
-    type: 'boolean',
-    defaultValue: false
-  })
-
-  options.matchPath = simpleSchema(options.matchPath, {
-    type: 'string',
-    defaultValue: ''
-  })
-
-  options.returnType = simpleSchema(options.returnType, {
-    type: 'string',
-    defaultValue: returnTypeEnums.FIRST_CLOSEST_MATCH,
-    enums: returnTypeEnums
-  })
-
-  options.thresholdType = simpleSchema(options.thresholdType, {
-    type: 'string',
-    defaultValue: SIMILARITY,
-    enums: thresholdTypeEnums
-  })
-
-  switch (options.thresholdType) {
-    case EDIT_DISTANCE:
-      options.threshold = simpleSchema(options.threshold, {
-        type: 'integer',
-        defaultValue: 20
+const runOptionsSchema = R.compose(
+  R.cond([
+    [
+      R.propEq('thresholdType', thresholdTypeEnums.EDIT_DISTANCE),
+      (options) => ({
+        ...options,
+        threshold: simpleSchema(options.threshold, {
+          type: 'integer',
+          default: 20
+        })
       })
-      break
-
-    case SIMILARITY:
-      options.threshold = simpleSchema(options.threshold, {
-        type: 'number',
-        defaultValue: 0.4
+    ],
+    [
+      R.propEq('thresholdType', thresholdTypeEnums.SIMILARITY),
+      (options) => ({
+        ...options,
+        threshold: simpleSchema(options.threshold, {
+          type: 'number',
+          default: 0.4
+        })
       })
-      break
-
-    /* istanbul ignore next */ // handled by simpleSchema
-    default:
-  }
-
-  options.trimSpace = simpleSchema(options.trimSpace, {
-    type: 'boolean',
-    defaultValue: false
+    ],
+    [
+      R.T,
+      () => {
+        throw new Error('unhandled `thresholdType`')
+      }
+    ]
+  ]),
+  (options = {}) => ({
+    caseSensitive: simpleSchema(options.caseSensitive, {
+      type: 'boolean',
+      default: false
+    }),
+    deburr: simpleSchema(options.deburr, {
+      type: 'boolean',
+      default: false
+    }),
+    matchPath: simpleSchema(options.matchPath, {
+      type: 'array',
+      default: []
+    }),
+    returnType: simpleSchema(options.returnType, {
+      type: 'string',
+      default: returnTypeEnums.FIRST_CLOSEST_MATCH,
+      enum: R.values(returnTypeEnums)
+    }),
+    threshold: options.threshold,
+    thresholdType: simpleSchema(options.thresholdType, {
+      type: 'string',
+      default: thresholdTypeEnums.SIMILARITY,
+      enum: R.values(thresholdTypeEnums)
+    }),
+    trimSpaces: simpleSchema(options.trimSpaces, {
+      type: 'boolean',
+      default: true
+    })
   })
+)
 
-  return options
-}
-
-module.exports = runOptionsSchema
+export default runOptionsSchema
